@@ -17,11 +17,16 @@ package com.linecorp.armeria.client.metric;
 
 import static java.util.Objects.requireNonNull;
 
+import java.util.function.BiPredicate;
 import java.util.function.Function;
+
+import javax.annotation.Nullable;
 
 import com.linecorp.armeria.client.HttpClient;
 import com.linecorp.armeria.common.HttpRequest;
 import com.linecorp.armeria.common.HttpResponse;
+import com.linecorp.armeria.common.RequestContext;
+import com.linecorp.armeria.common.logging.RequestLog;
 import com.linecorp.armeria.common.metric.MeterIdPrefixFunction;
 
 import io.micrometer.core.instrument.MeterRegistry;
@@ -50,10 +55,18 @@ public final class MetricCollectingClient extends AbstractMetricCollectingClient
     public static Function<? super HttpClient, MetricCollectingClient> newDecorator(
             MeterIdPrefixFunction meterIdPrefixFunction) {
         requireNonNull(meterIdPrefixFunction, "meterIdPrefixFunction");
-        return delegate -> new MetricCollectingClient(delegate, meterIdPrefixFunction);
+        return builder(meterIdPrefixFunction).newDecorator();
     }
 
-    MetricCollectingClient(HttpClient delegate, MeterIdPrefixFunction meterIdPrefixFunction) {
-        super(delegate, meterIdPrefixFunction);
+    /**
+     * Returns a new {@link MetricCollectingClientBuilder} instance.
+     */
+    public static MetricCollectingClientBuilder builder(MeterIdPrefixFunction meterIdPrefixFunction) {
+        return new MetricCollectingClientBuilder(meterIdPrefixFunction);
+    }
+
+    MetricCollectingClient(HttpClient delegate, MeterIdPrefixFunction meterIdPrefixFunction,
+                           @Nullable BiPredicate<? super RequestContext, ? super RequestLog> successFunction) {
+        super(delegate, meterIdPrefixFunction, successFunction);
     }
 }
