@@ -22,6 +22,7 @@ import static com.linecorp.armeria.common.metric.MoreMeters.newTimer;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
 import java.util.concurrent.atomic.LongAdder;
+import java.util.function.BiPredicate;
 
 import javax.annotation.Nullable;
 
@@ -50,8 +51,10 @@ public final class RequestMetricSupport {
     /**
      * Sets up request metrics.
      */
-    public static void setup(RequestContext ctx, AttributeKey<Boolean> requestMetricsSetKey,
-                             MeterIdPrefixFunction meterIdPrefixFunction, boolean server) {
+    public static void setup(
+            RequestContext ctx, AttributeKey<Boolean> requestMetricsSetKey,
+            MeterIdPrefixFunction meterIdPrefixFunction, boolean server,
+            @Nullable BiPredicate<? super RequestContext, ? super RequestLog> successFunction) {
         final Boolean isRequestMetricsSet = ctx.attr(requestMetricsSetKey);
 
         if (Boolean.TRUE.equals(isRequestMetricsSet)) {
@@ -64,10 +67,12 @@ public final class RequestMetricSupport {
                           RequestLogProperty.REQUEST_HEADERS,
                           RequestLogProperty.NAME,
                           RequestLogProperty.SESSION)
-           .thenAccept(log -> onRequest(log, meterIdPrefixFunction, server));
+           .thenAccept(log -> onRequest(log, meterIdPrefixFunction, server, successFunction));
     }
 
-    private static void onRequest(RequestLog log, MeterIdPrefixFunction meterIdPrefixFunction, boolean server) {
+    private static void onRequest(
+            RequestLog log, MeterIdPrefixFunction meterIdPrefixFunction, boolean server,
+            @Nullable BiPredicate<? super RequestContext, ? super RequestLog> successFunction) {
         final RequestContext ctx = log.context();
         final MeterRegistry registry = ctx.meterRegistry();
         final MeterIdPrefix activeRequestsId =
@@ -80,13 +85,14 @@ public final class RequestMetricSupport {
                                   new ActiveRequestMetrics(), ActiveRequestMetrics::doubleValue));
         activeRequestMetrics.increment();
         ctx.log().whenComplete().thenAccept(requestLog -> {
-            onResponse(requestLog, meterIdPrefixFunction, server);
+            onResponse(requestLog, meterIdPrefixFunction, server, successFunction);
             activeRequestMetrics.decrement();
         });
     }
 
-    private static void onResponse(RequestLog log, MeterIdPrefixFunction meterIdPrefixFunction,
-                                   boolean server) {
+    private static void onResponse(
+            RequestLog log, MeterIdPrefixFunction meterIdPrefixFunction, boolean server,
+            @Nullable BiPredicate<? super RequestContext, ? super RequestLog> successFunction) {
         final RequestContext ctx = log.context();
         final MeterRegistry registry = ctx.meterRegistry();
         final MeterIdPrefix idPrefix = meterIdPrefixFunction.completeRequestPrefix(registry, log);
@@ -95,7 +101,7 @@ public final class RequestMetricSupport {
             final ServiceRequestMetrics metrics = MicrometerUtil.register(registry, idPrefix,
                                                                           ServiceRequestMetrics.class,
                                                                           DefaultServiceRequestMetrics::new);
-            updateMetrics(log, metrics);
+            updateMetrics(ctx, log, metrics, successFunction);
             if (log.responseCause() instanceof RequestTimeoutException) {
                 metrics.requestTimeouts().increment();
             }
@@ -105,7 +111,7 @@ public final class RequestMetricSupport {
         final ClientRequestMetrics metrics = MicrometerUtil.register(registry, idPrefix,
                                                                      ClientRequestMetrics.class,
                                                                      DefaultClientRequestMetrics::new);
-        updateMetrics(log, metrics);
+        updateMetrics(ctx, log, metrics, successFunction);
         final ClientConnectionTimings timings = log.connectionTimings();
         if (timings != null) {
             metrics.connectionAcquisitionDuration().record(timings.connectionAcquisitionDurationNanos(),
@@ -141,7 +147,9 @@ public final class RequestMetricSupport {
         }
     }
 
-    private static void updateMetrics(RequestLog log, RequestMetrics metrics) {
+    private static void updateMetrics(
+            RequestContext ctx, RequestLog log, RequestMetrics metrics,
+            @Nullable BiPredicate<? super RequestContext, ? super RequestLog> successFunction) {
         if (log.requestCause() != null) {
             metrics.failure().increment();
             return;
@@ -153,7 +161,13 @@ public final class RequestMetricSupport {
         metrics.responseLength().record(log.responseLength());
         metrics.totalDuration().record(log.totalDurationNanos(), TimeUnit.NANOSECONDS);
 
-        if (isSuccess(log)) {
+        final boolean success;
+        if (successFunction != null) {
+            success = successFunction.test(ctx, log);
+        } else {
+            success = isSuccess(log);
+        }
+        if (success) {
             metrics.success().increment();
         } else {
             metrics.failure().increment();
