@@ -18,6 +18,7 @@ package com.linecorp.armeria.server;
 
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
+import static com.linecorp.armeria.server.RouteBuilder.FALLBACK_ROUTE;
 import static com.linecorp.armeria.server.RouteCache.wrapRouteDecoratingServiceRouter;
 import static com.linecorp.armeria.server.RouteCache.wrapVirtualHostRouter;
 import static java.util.Objects.requireNonNull;
@@ -34,6 +35,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.BiConsumer;
+import java.util.function.BiFunction;
 import java.util.function.Function;
 import java.util.stream.StreamSupport;
 
@@ -45,6 +47,9 @@ import org.slf4j.LoggerFactory;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
+
+import com.linecorp.armeria.server.RoutingTrie.Node;
+import com.linecorp.armeria.server.RoutingTrie.NodeProcessor;
 
 /**
  * A factory that creates a {@link Router} instance.
@@ -71,13 +76,43 @@ public final class Routers {
                             RejectedRouteHandler.class.getSimpleName(), e);
             }
         };
+
+        final Map<Route, ServiceConfig> newServiceConfigs = new HashMap<>();
+        final BiFunction<Route, ServiceConfig, ServiceConfig> fallbackValueConfigurator =
+                (originalRoute, fallbackServiceConfig) -> {
+                    final Route fallbackRoute = fallbackServiceConfig.route();
+
+                    if (originalRoute.complexity() == fallbackRoute.complexity() &&
+                        originalRoute.methods().containsAll(fallbackRoute.methods())) {
+                        return fallbackServiceConfig;
+                    }
+
+                    assert fallbackRoute.equals(FALLBACK_ROUTE)
+                            : "Fallback service must catch all requests, but its route is: " + fallbackRoute;
+                    final Route newRoute =
+                            originalRoute.toBuilder()
+                                         .pathMapping(CatchAllPathMapping.INSTANCE)
+                                         // Set this route as a fallback.
+                                         .fallback(true)
+                                         .build();
+                    // We have only one fallback ServiceConfig instance so finding a cached config
+                    // with a Route instance is okay at the moment.
+                    final ServiceConfig cachedConfig = newServiceConfigs.get(newRoute);
+                    if (cachedConfig != null) {
+                        return cachedConfig;
+                    }
+                    final ServiceConfig newConfig = fallbackServiceConfig.withRoute(newRoute);
+                    newServiceConfigs.put(newRoute, newConfig);
+                    return newConfig;
+                };
         final Set<Route> ambiguousRoutes =
                 resolveAmbiguousRoutes(StreamSupport.stream(configs.spliterator(), false)
                                                     .map(ServiceConfig::route)
                                                     .collect(toImmutableList()));
-        return wrapVirtualHostRouter(defaultRouter(configs, virtualHost.fallbackServiceConfig(),
-                                                   ServiceConfig::route, rejectionConsumer, false),
-                                     ambiguousRoutes);
+        return wrapVirtualHostRouter(
+                defaultRouter(configs, virtualHost.fallbackServiceConfig(), fallbackValueConfigurator,
+                              ServiceConfig::route, rejectionConsumer, false),
+                ambiguousRoutes);
     }
 
     /**
@@ -86,7 +121,7 @@ public final class Routers {
     public static Router<RouteDecoratingService> ofRouteDecoratingService(
             List<RouteDecoratingService> routeDecoratingServices) {
         return wrapRouteDecoratingServiceRouter(
-                sequentialRouter(routeDecoratingServices, null, RouteDecoratingService::route,
+                sequentialRouter(routeDecoratingServices, null, null, RouteDecoratingService::route,
                                  (route1, route2) -> {/* noop */}, true),
                 resolveAmbiguousRoutes(routeDecoratingServices.stream()
                                                               .map(RouteDecoratingService::route)
@@ -128,11 +163,12 @@ public final class Routers {
      * transformed to a {@link Router}.
      */
     private static <V> Router<V> defaultRouter(Iterable<V> values, @Nullable V fallbackValue,
+                                               @Nullable BiFunction<Route, V, V> fallbackValueConfigurator,
                                                Function<V, Route> routeResolver,
                                                BiConsumer<Route, Route> rejectionHandler,
                                                boolean isRouteDecorator) {
-        return new CompositeRouter<>(routers(values, fallbackValue, routeResolver, rejectionHandler,
-                                             isRouteDecorator),
+        return new CompositeRouter<>(routers(values, fallbackValue, fallbackValueConfigurator,
+                                             routeResolver, rejectionHandler, isRouteDecorator),
                                      Function.identity());
     }
 
@@ -141,6 +177,7 @@ public final class Routers {
      */
     @VisibleForTesting
     static <V> List<Router<V>> routers(Iterable<V> values, @Nullable V fallbackValue,
+                                       @Nullable BiFunction<Route, V, V> fallbackValueConfigurator,
                                        Function<V, Route> routeResolver,
                                        BiConsumer<Route, Route> rejectionHandler,
                                        boolean isRouteDecorator) {
@@ -162,13 +199,15 @@ public final class Routers {
 
             // Changed the router type.
             if (!group.isEmpty()) {
-                builder.add(router(addingTrie, group, fallbackValue, routeResolver, isRouteDecorator));
+                builder.add(router(addingTrie, group, fallbackValue, fallbackValueConfigurator,
+                                   routeResolver, isRouteDecorator));
             }
             addingTrie = !addingTrie;
             group.add(value);
         }
         if (!group.isEmpty()) {
-            builder.add(router(addingTrie, group, fallbackValue, routeResolver, isRouteDecorator));
+            builder.add(router(addingTrie, group, fallbackValue, fallbackValueConfigurator,
+                               routeResolver, isRouteDecorator));
         }
         return builder.build();
     }
@@ -177,12 +216,13 @@ public final class Routers {
      * Returns only the sequential implementation of {@link Router}.
      */
     private static <V> Router<V> sequentialRouter(Iterable<V> values, @Nullable V fallbackValue,
+                                                  @Nullable BiFunction<Route, V, V> fallbackValueConfigurator,
                                                   Function<V, Route> routeResolver,
                                                   BiConsumer<Route, Route> rejectionHandler,
                                                   boolean isRouteDecorator) {
         rejectDuplicateMapping(values, routeResolver, rejectionHandler);
-        return router(false, Lists.newArrayList(values), fallbackValue, routeResolver,
-                      isRouteDecorator);
+        return router(false, Lists.newArrayList(values), fallbackValue, fallbackValueConfigurator,
+                      routeResolver, isRouteDecorator);
     }
 
     private static <V> void rejectDuplicateMapping(
@@ -238,6 +278,7 @@ public final class Routers {
      * Returns a {@link Router} implementation which is using one of {@link RoutingTrie} and {@link List}.
      */
     private static <V> Router<V> router(boolean isTrie, List<V> values, @Nullable V fallbackValue,
+                                        @Nullable BiFunction<Route, V, V> fallbackValueConfigurator,
                                         Function<V, Route> routeResolver, boolean isRouteDecorator) {
         final Comparator<V> valueComparator =
                 Comparator.comparingInt(e -> -1 * routeResolver.apply(e).complexity());
@@ -257,8 +298,11 @@ public final class Routers {
                     final String path = route.paths().get(0);
                     final int pathLen = path.length();
                     if (pathLen > 1 && path.charAt(pathLen - 1) == '/') {
+                        final V newFallbackValue =
+                                fallbackValueConfigurator != null ?
+                                fallbackValueConfigurator.apply(route, fallbackValue) : fallbackValue;
                         builder.add(path.substring(0, pathLen - 1),
-                                    fallbackValue, /* hasHighPrecedence */ false);
+                                    newFallbackValue, /* hasHighPrecedence */ false);
                     }
                 }
             }
@@ -283,65 +327,66 @@ public final class Routers {
     /**
      * Finds the most suitable service from the given {@link ServiceConfig} list.
      */
-    private static <V> Routed<V> findBest(RoutingContext routingCtx, @Nullable List<V> values,
-                                          Function<V, Route> routeResolver) {
+    private static <V> Routed<V> findBest(List<Routed<V>> routes) {
         Routed<V> result = Routed.empty();
-        if (values != null) {
-            for (V value : values) {
-                final Route route = routeResolver.apply(value);
-                final RoutingResult routingResult = route.apply(routingCtx, false);
-                if (routingResult.isPresent()) {
-                    //
-                    // The services are sorted as follows:
-                    //
-                    // 1) the service with method and media type negotiation
-                    //    (consumable and producible)
-                    // 2) the service with method and producible media type negotiation
-                    // 3) the service with method and consumable media type negotiation
-                    // 4) the service with method negotiation
-                    // 5) the other services (in a registered order)
-                    //
-                    // 1) and 2) may produce a score between the lowest and the highest because they should
-                    // negotiate the produce type with the value of 'Accept' header.
-                    // 3), 4) and 5) always produces the lowest score.
-                    //
+        for (Routed<V> route : routes) {
+            final RoutingResult routingResult = route.routingResult();
+            if (routingResult.isPresent()) {
+                //
+                // The services are sorted as follows:
+                //
+                // 1) the service with method and media type negotiation
+                //    (consumable and producible)
+                // 2) the service with method and producible media type negotiation
+                // 3) the service with method and consumable media type negotiation
+                // 4) the service with method negotiation
+                // 5) the other services (in a registered order)
+                //
+                // 1) and 2) may produce a score between the lowest and the highest because they should
+                // negotiate the produce type with the value of 'Accept' header.
+                // 3), 4) and 5) always produces the lowest score.
+                //
 
-                    // Found the best matching.
-                    if (routingResult.hasHighestScore()) {
-                        result = Routed.of(route, routingResult, value);
-                        break;
-                    }
+                // Found the best matching.
+                if (routingResult.hasHighestScore()) {
+                    result = route;
+                    break;
+                }
 
-                    // We have still a chance to find a better matching.
-                    if (result.isPresent()) {
-                        if (routingResult.score() > result.routingResult().score()) {
-                            // Replace the candidate with the new one only if the score is better.
-                            // If the score is same, we respect the order of service registration.
-                            result = Routed.of(route, routingResult, value);
-                        }
-                    } else {
-                        // Keep the result as a candidate.
-                        result = Routed.of(route, routingResult, value);
+                // We have still a chance to find a better matching.
+                if (result.isPresent()) {
+                    if (routingResult.score() > result.routingResult().score()) {
+                        // Replace the candidate with the new one only if the score is better.
+                        // If the score is same, we respect the order of service registration.
+                        result = route;
                     }
+                } else {
+                    // Keep the result as a candidate.
+                    result = route;
                 }
             }
         }
         return result;
     }
 
-    private static <V> List<Routed<V>> findAll(RoutingContext routingCtx, List<V> values,
-                                               Function<V, Route> routeResolver,
-                                               boolean isRouteDecorator) {
-        final ImmutableList.Builder<Routed<V>> builder = ImmutableList.builderWithExpectedSize(values.size());
+    private static <V> List<Routed<V>> getRouteCandidates(RoutingContext routingCtx, List<V> values,
+                                                          Function<V, Route> routeResolver,
+                                                          boolean isRouteDecorator) {
+        ImmutableList.Builder<Routed<V>> builder = null;
+        int remaining = values.size();
 
         for (V value : values) {
             final Route route = routeResolver.apply(value);
             final RoutingResult routingResult = route.apply(routingCtx, isRouteDecorator);
             if (routingResult.isPresent()) {
+                if (builder == null) {
+                    builder = ImmutableList.builderWithExpectedSize(remaining);
+                }
                 builder.add(Routed.of(route, routingResult, value));
             }
+            remaining--;
         }
-        return builder.build();
+        return builder != null ? builder.build() : ImmutableList.of();
     }
 
     private static final class TrieRouter<V> implements Router<V> {
@@ -358,18 +403,51 @@ public final class Routers {
 
         @Override
         public Routed<V> find(RoutingContext routingCtx) {
-            return findBest(routingCtx, trie.find(routingCtx.path()), routeResolver);
+            final RouteCandidateCollectingNodeProcessor processor =
+                    new RouteCandidateCollectingNodeProcessor(routingCtx);
+            trie.find(routingCtx.path(), processor);
+            return findBest(processor.collectRouteCandidates());
         }
 
         @Override
         public List<Routed<V>> findAll(RoutingContext routingCtx) {
-            return Routers.findAll(routingCtx, trie.findAll(routingCtx.path()),
-                                   routeResolver, isRouteDecorator);
+            return getRouteCandidates(routingCtx, trie.findAll(routingCtx.path()),
+                                      routeResolver, isRouteDecorator);
         }
 
         @Override
         public void dump(OutputStream output) {
             trie.dump(output);
+        }
+
+        private final class RouteCandidateCollectingNodeProcessor implements NodeProcessor<V> {
+            private final RoutingContext routingCtx;
+            @Nullable
+            private ImmutableList.Builder<Routed<V>> routeCollector;
+
+            private RouteCandidateCollectingNodeProcessor(RoutingContext routingCtx) {
+                this.routingCtx = routingCtx;
+            }
+
+            @Nullable
+            @Override
+            public Node<V> process(Node<V> node) {
+                final List<Routed<V>> list =
+                        getRouteCandidates(routingCtx, node.values, routeResolver, isRouteDecorator);
+                if (list.isEmpty()) {
+                    // Not acceptable node.
+                    return null;
+                }
+                if (routeCollector == null) {
+                    routeCollector = ImmutableList.builder();
+                }
+                routeCollector.addAll(list);
+                return node;
+            }
+
+            List<Routed<V>> collectRouteCandidates() {
+                return routeCollector != null ? routeCollector.build() : ImmutableList.of();
+            }
         }
     }
 
@@ -387,12 +465,12 @@ public final class Routers {
 
         @Override
         public Routed<V> find(RoutingContext routingCtx) {
-            return findBest(routingCtx, values, routeResolver);
+            return findBest(getRouteCandidates(routingCtx, values, routeResolver, false));
         }
 
         @Override
         public List<Routed<V>> findAll(RoutingContext routingCtx) {
-            return Routers.findAll(routingCtx, values, routeResolver, isRouteDecorator);
+            return getRouteCandidates(routingCtx, values, routeResolver, isRouteDecorator);
         }
 
         @Override
