@@ -14,8 +14,9 @@
  * under the License.
  */
 
+import JSONbig from 'json-bigint';
 import { Endpoint, Method } from '../specification';
-import prettify from '../json-prettify';
+import jsonPrettify from '../json-prettify';
 
 import Transport from './transport';
 
@@ -112,16 +113,17 @@ export default class AnnotatedHttpTransport extends Transport {
       body: bodyJson,
     });
     const applicationType = httpResponse.headers.get('content-type') || '';
-    const response = await httpResponse.text();
-    if (response.length > 0) {
-      if (applicationType.indexOf('json') > -1) {
-        const prettified = prettify(response);
-        if (prettified.length === 0) {
-          return response;
-        }
+    if (applicationType.indexOf('json') >= 0) {
+      const json = JSONbig.parse(await httpResponse.text());
+      const prettified = jsonPrettify(JSONbig.stringify(json));
+      if (prettified.length > 0) {
         return prettified;
       }
-      return response;
+    }
+
+    const responseText = await httpResponse.text();
+    if (responseText.length > 0) {
+      return responseText;
     }
     return '<zero-length response>';
   }
