@@ -21,14 +21,17 @@ import com.linecorp.armeria.common.grpc.GrpcJsonMarshaller
 import com.linecorp.armeria.common.scalapb.ScalaPbJsonMarshaller.{
   jsonDefaultParser,
   jsonDefaultPrinter,
-  messageCompanionCache
+  messageCompanionCache,
+  typeMapperMethodCache
 }
 import io.grpc.MethodDescriptor.Marshaller
 import java.io.{InputStream, OutputStream}
-import java.util.concurrent.ConcurrentMap
+import java.lang.invoke.{MethodHandle, MethodHandles}
+import java.util.concurrent.{ConcurrentHashMap, ConcurrentMap}
 import scala.io.{Codec, Source}
+import scalapb.grpc.TypeMappedMarshaller
 import scalapb.json4s.{Parser, Printer}
-import scalapb.{GeneratedMessage, GeneratedMessageCompanion, GeneratedSealedOneof}
+import scalapb.{GeneratedMessage, GeneratedMessageCompanion, GeneratedSealedOneof, TypeMapper}
 
 /**
  * A [[com.linecorp.armeria.common.grpc.GrpcJsonMarshaller]] that serializes and deserializes
@@ -54,7 +57,20 @@ final class ScalaPbJsonMarshaller private (
     val companion = getMessageCompanion(marshaller)
     val jsonString = Source.fromInputStream(in)(Codec.UTF8).mkString
     val message = jsonParser.fromJsonString(jsonString)(companion)
-    message.asInstanceOf[A]
+    marshaller match {
+      case marshaller: TypeMappedMarshaller[_, _] =>
+        val method = typeMapperMethodCache.computeIfAbsent(
+          marshaller,
+          key => {
+            val field = key.getClass.getDeclaredField("typeMapper")
+            field.setAccessible(true)
+            MethodHandles.lookup().unreflectGetter(field)
+          })
+        val typeMapper = method.invoke(marshaller).asInstanceOf[TypeMapper[GeneratedMessage, A]]
+        typeMapper.toCustom(message)
+      case _ =>
+        message.asInstanceOf[A]
+    }
   }
 
   private def getMessageCompanion[A](marshaller: Marshaller[A]): GeneratedMessageCompanion[GeneratedMessage] = {
@@ -79,6 +95,9 @@ final class ScalaPbJsonMarshaller private (
 object ScalaPbJsonMarshaller {
 
   private val messageCompanionCache: ConcurrentMap[Marshaller[_], GeneratedMessageCompanion[GeneratedMessage]] =
+    new MapMaker().weakKeys().makeMap()
+
+  private val typeMapperMethodCache: ConcurrentMap[Marshaller[_], MethodHandle] =
     new MapMaker().weakKeys().makeMap()
 
   private val jsonDefaultPrinter: Printer = new Printer().includingDefaultValueFields
