@@ -17,7 +17,7 @@
 package com.linecorp.armeria.server.scalapb
 
 import com.fasterxml.jackson.databind.{JsonNode, ObjectMapper}
-import com.google.common.collect._
+import com.google.common.collect.{ImmutableList, ImmutableMap, ImmutableSet, MapMaker}
 import com.google.protobuf.CodedInputStream
 import com.linecorp.armeria.common.AggregatedHttpRequest
 import com.linecorp.armeria.common.annotation.UnstableApi
@@ -32,7 +32,7 @@ import java.nio.charset.StandardCharsets
 import java.util.concurrent.ConcurrentMap
 import javax.annotation.Nullable
 import scalapb.json4s.Parser
-import scalapb.{GeneratedMessage, GeneratedMessageCompanion}
+import scalapb.{GeneratedMessage, GeneratedMessageCompanion, GeneratedSealedOneof}
 
 /**
  * A [[com.linecorp.armeria.server.annotation.RequestConverterFunction]] which converts either
@@ -68,6 +68,9 @@ object ScalaPbRequestConverterFunction {
   private val methodCache: ConcurrentMap[Class[_], MethodHandle] =
     new MapMaker().weakKeys.makeMap()
 
+  private val toOneofMethodCache: ConcurrentMap[Class[_], MethodHandle] =
+    new MapMaker().weakKeys.makeMap()
+
   private val defaultInstanceCache: ConcurrentMap[Class[_], GeneratedMessage] =
     new MapMaker().weakKeys.makeMap()
 
@@ -94,9 +97,10 @@ object ScalaPbRequestConverterFunction {
       clazz,
       key =>
         try {
+          val genClass = extractGeneratedMessageType(key)
           val lookup = MethodHandles.publicLookup()
-          val mt = MethodType.methodType(key)
-          val methodHandle = lookup.findStatic(key, "defaultInstance", mt)
+          val mt = MethodType.methodType(genClass)
+          val methodHandle = lookup.findStatic(genClass, "defaultInstance", mt)
           methodHandle.invoke()
         } catch {
           case _: NoSuchMethodError | _: IllegalAccessException =>
@@ -116,9 +120,10 @@ object ScalaPbRequestConverterFunction {
       clazz,
       key =>
         try {
+          val genClass = extractGeneratedMessageType(key)
           val lookup = MethodHandles.publicLookup()
-          val mt = MethodType.methodType(key, key, classOf[CodedInputStream])
-          lookup.findStatic(key, "merge", mt)
+          val mt = MethodType.methodType(genClass, genClass, classOf[CodedInputStream])
+          lookup.findStatic(genClass, "merge", mt)
         } catch {
           case _: NoSuchFieldException | _: ClassNotFoundException =>
             unknownMethodHandle
@@ -129,6 +134,51 @@ object ScalaPbRequestConverterFunction {
       throw new IllegalStateException(s"Failed to find a static merge method from $clazz")
     methodHandle
   }
+
+  /**
+   * Extracts a [[scalapb.GeneratedMessage]] class from the specified [[scalapb.GeneratedSealedOneof]] or
+   * returns itself.
+   *
+   * @param clazz Either `Class[GeneratedMessage]` or `Class[GeneratedSealedOneof]`.
+   */
+  private def extractGeneratedMessageType(clazz: Class[_]): Class[GeneratedMessage] = {
+    val isOneOf = classOf[GeneratedSealedOneof].isAssignableFrom(clazz)
+    val messageType =
+      if (isOneOf)
+        clazz.getDeclaredMethod("asMessage").getReturnType
+      else
+        clazz
+    messageType.asInstanceOf[Class[GeneratedMessage]]
+  }
+
+  /**
+   * Converts a [[scalapb.GeneratedMessage]] into a [[scalapb.GeneratedSealedOneof]]
+   * if the `expectedResultType` is a subtype of [[scalapb.GeneratedSealedOneof]].
+   */
+  private def toGenerateMessageOrOneof(
+      expectedResultType: Class[_],
+      result: GeneratedMessage): Any with Serializable =
+    if (classOf[GeneratedSealedOneof].isAssignableFrom(expectedResultType)) {
+      val methodHandle = toOneofMethodCache.computeIfAbsent(
+        expectedResultType,
+        key =>
+          try {
+            val lookup = MethodHandles.publicLookup()
+            val mt = MethodType.methodType(key)
+            lookup.findVirtual(result.getClass, s"to${key.getSimpleName}", mt)
+          } catch {
+            case _: NoSuchFieldException | _: ClassNotFoundException =>
+              unknownMethodHandle
+          }
+      )
+
+      if (methodHandle eq unknownMethodHandle)
+        throw new IllegalStateException(
+          s"Failed to find to${expectedResultType.getSimpleName} method from $result")
+
+      methodHandle.invoke(result)
+    } else
+      result
 
   /**
    * Returns a companion object used to convert JSON to a [[scalapb.GeneratedMessage]].
@@ -180,18 +230,21 @@ final class ScalaPbRequestConverterFunction private (jsonParser: Parser, resultT
       else contentType.charset(StandardCharsets.UTF_8)
 
     if (resultType == ResultType.PROTOBUF ||
-      (resultType == ResultType.UNKNOWN &&
-      classOf[GeneratedMessage].isAssignableFrom(expectedResultType))) {
+      (resultType == ResultType.UNKNOWN && isProtobufMessage(expectedResultType))) {
       val mergeMH = getMergeMethod(expectedResultType)
       if (contentType == null || isProtobuf(contentType)) {
         val is = request.content.toInputStream
-        try return mergeMH.invoke(getDefaultInstance(expectedResultType), CodedInputStream.newInstance(is))
-        finally if (is != null)
+        try {
+          val message = mergeMH
+            .invoke(getDefaultInstance(expectedResultType), CodedInputStream.newInstance(is))
+            .asInstanceOf[GeneratedMessage]
+          return toGenerateMessageOrOneof(expectedResultType, message).asInstanceOf[Object]
+        } finally if (is != null)
           is.close()
       }
       if (isJson(contentType)) {
         val jsonString = request.content(charset)
-        return toJsonGeneratedMessage(expectedResultType, jsonString).asInstanceOf[Object]
+        return jsonToScalaPbMessage(expectedResultType, jsonString).asInstanceOf[Object]
       }
 
       if (!isJson(contentType) || expectedParameterizedResultType == null)
@@ -223,27 +276,26 @@ final class ScalaPbRequestConverterFunction private (jsonParser: Parser, resultT
       val iter = jsonNode.iterator()
       resultType match {
         case LIST_PROTOBUF | SET_PROTOBUF =>
-          val builder = {
-            if (resultType == LIST_PROTOBUF) ImmutableList.builderWithExpectedSize[GeneratedMessage](size)
-            else ImmutableSet.builderWithExpectedSize[GeneratedMessage](size)
-          }
+          val builder =
+            if (resultType == LIST_PROTOBUF) ImmutableList.builderWithExpectedSize[Any with Serializable](size)
+            else ImmutableSet.builderWithExpectedSize[Any with Serializable](size)
 
           while (iter.hasNext)
-            builder.add(toJsonGeneratedMessage(messageType, iter.next()))
+            builder.add(jsonToScalaPbMessage(messageType, iter.next()))
           builder.build
 
         case SCALA_LIST_PROTOBUF | SCALA_VECTOR_PROTOBUF | SCALA_SET_PROTOBUF =>
           val builder =
             if (resultType == SCALA_LIST_PROTOBUF)
-              List.newBuilder[GeneratedMessage]
+              List.newBuilder[Any]
             else if (resultType == SCALA_VECTOR_PROTOBUF)
-              Vector.newBuilder[GeneratedMessage]
+              Vector.newBuilder[Any]
             else
-              Set.newBuilder[GeneratedMessage]
+              Set.newBuilder[Any]
           builder.sizeHint(size)
 
           while (iter.hasNext)
-            builder += toJsonGeneratedMessage(messageType, iter.next())
+            builder += jsonToScalaPbMessage(messageType, iter.next())
           builder.result()
 
         case _ => RequestConverterFunction.fallthrough
@@ -253,20 +305,20 @@ final class ScalaPbRequestConverterFunction private (jsonParser: Parser, resultT
       val iter = jsonNode.fields
       resultType match {
         case MAP_PROTOBUF =>
-          val builder = ImmutableMap.builderWithExpectedSize[String, GeneratedMessage](size)
+          val builder = ImmutableMap.builderWithExpectedSize[String, Any](size)
 
           while (iter.hasNext) {
             val entry = iter.next
-            builder.put(entry.getKey, toJsonGeneratedMessage(messageType, entry.getValue))
+            builder.put(entry.getKey, jsonToScalaPbMessage(messageType, entry.getValue))
           }
           builder.build
         case SCALA_MAP_PROTOBUF =>
-          val builder = Map.newBuilder[String, GeneratedMessage]
+          val builder = Map.newBuilder[String, Any]
           builder.sizeHint(size)
 
           while (iter.hasNext) {
             val entry = iter.next
-            builder += entry.getKey -> toJsonGeneratedMessage(messageType, entry.getValue)
+            builder += entry.getKey -> jsonToScalaPbMessage(messageType, entry.getValue)
           }
           builder.result()
 
@@ -276,11 +328,21 @@ final class ScalaPbRequestConverterFunction private (jsonParser: Parser, resultT
       RequestConverterFunction.fallthrough
   }
 
-  private def toJsonGeneratedMessage(expectedResultType: Class[_], node: JsonNode): GeneratedMessage = {
+  /**
+   * Returns a deserialized [[scalapb.GeneratedMessage]] or [[scalapb.GeneratedSealedOneof]] from the
+   * `JsonNode`.
+   */
+  private def jsonToScalaPbMessage(expectedResultType: Class[_], node: JsonNode): Any with Serializable = {
     val json = mapper.writeValueAsString(node)
-    toJsonGeneratedMessage(expectedResultType, json)
+    jsonToScalaPbMessage(expectedResultType, json)
   }
 
-  private def toJsonGeneratedMessage(expectedResultType: Class[_], json: String): GeneratedMessage =
-    jsonParser.fromJsonString(json)(getCompanion(expectedResultType))
+  /**
+   * Returns a deserialized [[scalapb.GeneratedMessage]] or [[scalapb.GeneratedSealedOneof]] from the `json`.
+   */
+  private def jsonToScalaPbMessage(expectedResultType: Class[_], json: String): Any with Serializable = {
+    val messageType = extractGeneratedMessageType(expectedResultType)
+    val message: GeneratedMessage = jsonParser.fromJsonString(json)(getCompanion(messageType))
+    toGenerateMessageOrOneof(expectedResultType, message)
+  }
 }
